Refactor spawning items/enemies. They're now broken into 3 functions each: - One for spawning from map data. - One for spawning from any array of updates. - One for spawning a single thing.
These feed into each other. For isntance, spawnEnemiesFromMap() will generate an array of enemy updates which is sent to spawnEnemies(), which in turn spawns each enemy with the spawnEnemy() method.

This will be used in the near future to enable spawning based on map data, or based on a batch update from the authoritative player in MP, or from a single update if the recipient client notices that the item to be updated does not exist.