Add all images, and image packing script that combines into one texture file.
This is all animations or static images for all sprites. The tilesets are still separate.

Please note the images have been slightly renamed from their original names when packed separately. This means the code will need to be tweaked for the new names before the packed texture file can be used. So right now all of these files and the packing script are extra and not used. The renaming does make the names more concise and more consistent.